Split Inkling sconv accuracy tests by reference case
